+++ src/components/Card/ProjectCard/index.ts
@@ -1,1 +1,1 @@
-export * from "./card"
+export * from "./ProjectCard";
